Fixed folder structure.
