+++ BudgetManager/BudgetManager.csproj
@@ -20,8 +20,4 @@
     <PackageReference Include="Swashbuckle.AspNetCore" Version="6.7.3" />
   </ItemGroup>
 
-  <ItemGroup>
-    <Folder Include="Handlers\Transactions\Queries\" />
-  </ItemGroup>
-
 </Project>
