+++ BudgetManager/Features/Transactions/Commands/SaveTransactionCommand.cs
@@ -1,5 +1,7 @@
 ﻿using BudgetManager.Dto.Transaction;
+using BudgetManager.Mappers;
 using BudgetManager.Models;
+using BudgetManager.Services;
 using MediatR;
 
 namespace BudgetManager.Features.Transactions.Commands
@@ -19,6 +21,25 @@ namespace BudgetManager.Features.Transactions.Commands
             Date = date;
             Price = price;
             Category = category;
+        }
+    }
+    public class SaveTransactionCommandHandler : IRequestHandler<SaveTransactionCommand, TransactionDto>
+    {
+        private readonly ITransactionService _transactionService;
+        private readonly ITransactionMapper _transactionMapper;
+
+        public SaveTransactionCommandHandler(ITransactionService transactionService, ITransactionMapper transactionMapper)
+        {
+            _transactionService = transactionService;
+            _transactionMapper = transactionMapper;
+        }
+
+        public async Task<TransactionDto> Handle(SaveTransactionCommand request, CancellationToken cancellationToken)
+        {
+            if (request == null)
+                throw new ArgumentNullException(nameof(request));
+
+            return await _transactionService.AddTransaction(_transactionMapper.Map(request));
         }
     }
 }
